brain/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
websockets==14.1
httpx>=0.27.0,<0.29.0
ollama>=0.4.4
chromadb==0.5.23
faster-whisper==1.2.1
sounddevice==0.5.5
numpy>=1.26.0
kokoro-onnx==0.5.0
python-dotenv==1.0.1
pyyaml==6.0.2
psutil>=5.9.0
cryptography>=42.0.0
pypykatz>=0.6.9
impacket>=0.12.0
pywin32>=306; sys_platform == "win32"
